Create Conversation with a Computer
This is one of the first scripts I ever wrote. It gave me the opportunity to learn about how to handle strings. There is concatenation, storing user input as a variable, as well as printing output to the screen. Since we could make the topic of conversation anything we wanted, I choose to make a fun interaction with a magical movie teller.